refactored tests: grouper, ilen, intersperse
