+++ tests/test_itertools_cfuncs.py
@@ -14,258 +14,6 @@ from helper_leak import memory_leak
 from helper_pytest_monkeypatch import pytest_raises
 
 
-def test_grouper():
-    grouper = iteration_utilities.grouper
-    # Empty iterable
-    assert list(grouper([], 2)) == []
-
-    # no fillvalue + truncate
-    assert list(grouper([1], 3)) == [(1, )]
-    assert list(grouper([1, 2], 3)) == [(1, 2)]
-    assert list(grouper([1, 2, 3], 3)) == [(1, 2, 3)]
-    assert list(grouper([1, 2, 3, 4], 3)) == [(1, 2, 3), (4, )]
-    assert list(grouper([1, 2, 3, 4, 5], 3)) == [(1, 2, 3), (4, 5)]
-    assert list(grouper([1, 2, 3, 4, 5, 6], 3)) == [(1, 2, 3), (4, 5, 6)]
-
-    # with fillvalue
-    assert list(grouper([1], 3,
-                        fillvalue=0)) == [(1, 0, 0)]
-    assert list(grouper([1, 2], 3,
-                        fillvalue=0)) == [(1, 2, 0)]
-    assert list(grouper([1, 2, 3], 3,
-                        fillvalue=0)) == [(1, 2, 3)]
-    assert list(grouper([1, 2, 3, 4], 3,
-                        fillvalue=0)) == [(1, 2, 3), (4, 0, 0)]
-    assert list(grouper([1, 2, 3, 4, 5], 3,
-                        fillvalue=0)) == [(1, 2, 3), (4, 5, 0)]
-    assert list(grouper([1, 2, 3, 4, 5, 6], 3,
-                        fillvalue=0)) == [(1, 2, 3), (4, 5, 6)]
-
-    # with truncate
-    assert list(grouper([1], 3,
-                        truncate=True)) == []
-    assert list(grouper([1, 2], 3,
-                        truncate=True)) == []
-    assert list(grouper([1, 2, 3], 3,
-                        truncate=True)) == [(1, 2, 3)]
-    assert list(grouper([1, 2, 3, 4], 3,
-                        truncate=True)) == [(1, 2, 3)]
-    assert list(grouper([1, 2, 3, 4, 5], 3,
-                        truncate=True)) == [(1, 2, 3)]
-    assert list(grouper([1, 2, 3, 4, 5, 6], 3,
-                        truncate=True)) == [(1, 2, 3), (4, 5, 6)]
-
-    # failures
-    with pytest.raises(TypeError):  # fillvalue + truncate is forbidden
-        grouper([1, 2, 3], 2, fillvalue=None, truncate=True)
-
-    with pytest.raises(ValueError):  # n must be > 0
-        grouper([1, 2, 3], 0)
-
-    with pytest.raises(TypeError):  # iterable must be iterable
-        grouper(1, 2)
-
-
-def test_grouper_memoryleak():
-    grouper = iteration_utilities.grouper
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-    # Empty iterable
-    def test():
-        list(grouper([], 2))
-    assert not memory_leak(test)
-
-    # no fillvalue + truncate
-    def test():
-        list(grouper([Test(1)], 3))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2)], 3))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3)], 3))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4)], 3))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4), Test(5)], 3))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4), Test(5), Test(6)],
-                     3))
-    assert not memory_leak(test)
-
-    # with fillvalue
-    def test():
-        list(grouper([Test(1)], 3,
-                     fillvalue=Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2)], 3,
-                     fillvalue=Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3)], 3,
-                     fillvalue=Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4)], 3,
-                     fillvalue=Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4), Test(5)], 3,
-                     fillvalue=Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4), Test(5), Test(6)], 3,
-                     fillvalue=Test(0)))
-    assert not memory_leak(test)
-
-    # with truncate
-    def test():
-        list(grouper([Test(1)], 3,
-                     truncate=True))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2)], 3,
-                     truncate=True))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3)], 3,
-                     truncate=True))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4)], 3,
-                     truncate=True))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4), Test(5)], 3,
-                     truncate=True))
-    assert not memory_leak(test)
-
-    def test():
-        list(grouper([Test(1), Test(2), Test(3), Test(4), Test(5), Test(6)], 3,
-                     truncate=True))
-    assert not memory_leak(test)
-
-    # failures
-    def test():
-        with pytest_raises(TypeError):  # fillvalue + truncate is forbidden
-            grouper([Test(1), Test(2), Test(3)], 2,
-                    fillvalue=Test(0), truncate=True)
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(ValueError):  # n must be > 0
-            grouper([Test(1), Test(2), Test(3)], 0)
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):  # iterable must be iterable
-            grouper(Test(1), 2)
-    assert not memory_leak(test)
-
-
-def test_ilen():
-    ilen = iteration_utilities.ilen
-
-    assert ilen([]) == 0
-    assert ilen(range(10)) == 10
-    assert ilen([1, 2, 3, 4, 5]) == 5
-
-    generator = (i for i in [1, 2, 3, 4, 5])
-    assert ilen(generator) == 5
-    assert ilen(generator) == 0
-
-    with pytest.raises(TypeError):
-        ilen(100)
-
-
-def test_ilen_memoryleak():
-    ilen = iteration_utilities.ilen
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-    def test():
-        ilen([])
-    assert not memory_leak(test)
-
-    def test():
-        ilen(range(10))
-    assert not memory_leak(test)
-
-    def test():
-        ilen([Test(1), Test(2), Test(3), Test(4), Test(5)])
-    assert not memory_leak(test)
-
-    def test():
-        generator = (i for i in [Test(1), Test(2), Test(3), Test(4), Test(5)])
-        ilen(generator) == 5
-        ilen(generator) == 0
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):
-            ilen(100)
-    assert not memory_leak(test)
-
-
-def test_intersperse():
-    intersperse = iteration_utilities.intersperse
-
-    assert list(intersperse([], 0)) == []
-    assert list(intersperse([1], 0)) == [1]
-    assert list(intersperse([1, 2], 0)) == [1, 0, 2]
-
-    with pytest.raises(TypeError):
-        intersperse(100, 0)
-
-
-def test_intersperse_memoryleak():
-    intersperse = iteration_utilities.intersperse
-
-    class Test(object):
-        def __init__(self, value):
-            self.value = value
-
-    def test():
-        list(intersperse([], Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        list(intersperse([Test(1)], Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        list(intersperse([Test(1), Test(2)], Test(0)))
-    assert not memory_leak(test)
-
-    def test():
-        with pytest_raises(TypeError):
-            intersperse(Test(100), Test(0))
-    assert not memory_leak(test)
-
-
 def test_complement():
     complement = iteration_utilities.complement
